Hustlr-websockets/requirements.txt
asgiref==3.11.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
cbor2==5.8.0
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
constantly==23.10.4
crispy-bootstrap4==2025.6
cryptography==46.0.4
daphne==4.2.1
Django==5.2.1
django-cors-headers==4.9.0
django-crispy-forms==2.4
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
msgpack==1.1.2
openpyxl==3.1.5
packaging==26.0
pillow==11.3.0
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.10.1
pyOpenSSL==25.3.0
PyQt6==6.10.2
PyQt6-Qt6==6.10.1
PyQt6_sip==13.11.0
python-dotenv==1.2.1
pytz==2025.2
redis==7.1.1
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
zope.interface==8.2
